Improve commandline options: Add -w option, use better default values
A separate -w option in *addition* to the previous -t option allows to customize
the target column that the frame rows are aligned/shifted to.

In most cases specifying -w is sufficient, since the -t option will default to the same
value as given to -w (resulting in the frame rows to be aligned towards the horizontal
center of the video frames).